requirements.txt
streamlit
joblib
scikit-learn
pandas
numpy
matplotlib
seaborn
joblib
google-generativeai
